+++ app/src/main/java/joshuatee/wx/util/UtilityNetworkIO.kt
@@ -34,6 +34,11 @@ import okhttp3.Request
 
 object UtilityNetworkIO {
 
+    private const val USER_AGENT =
+        "Android ${GlobalVariables.PACKAGE_NAME} ${GlobalVariables.EMAIL}"
+    private const val ACCEPT_HEADER =
+        "text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.9"
+
     private fun getStringFromUrlNew(url: String, withNewLine: Boolean): String {
         UtilityLog.download("getStringFromUrlNew $withNewLine: $url")
         val out = StringBuilder(5000)
@@ -106,5 +111,142 @@ object UtilityNetworkIO {
     } catch (e: IOException) {
         UtilityLog.handleException(e)
         null
+    }
+
+    // used for CapAlert (XML)
+    fun getStringFromUrlXml(url: String): String {
+        UtilityLog.download("getStringFromURLBase: $url")
+        val out = StringBuilder(5000)
+        try {
+            val request = Request.Builder()
+                .url(url)
+                .header("User-Agent", USER_AGENT)
+                .addHeader("Accept", "application/atom+xml")
+                .build()
+            val response = MyApplication.httpClient.newCall(request).execute()
+            val inputStream = BufferedInputStream(response.body.byteStream())
+            val bufferedReader = BufferedReader(InputStreamReader(inputStream))
+            var line: String? = bufferedReader.readLine()
+            while (line != null) {
+                out.append(line)
+                line = bufferedReader.readLine()
+            }
+            bufferedReader.close()
+        } catch (e: Exception) {
+            UtilityLog.handleException(e)
+        }
+        return out.toString()
+    }
+
+    // target for String.getNwsHtml()
+    fun getStringFromUrlBaseNoAcceptHeader1(url: String): String {
+        UtilityLog.download("getStringFromUrlBaseNoAcceptHeader1 getNwsHtml: $url")
+        val out = StringBuilder(5000)
+        try {
+            val request = Request.Builder()
+                .url(url)
+                .header("User-Agent", USER_AGENT)
+                .build()
+            val response = MyApplication.httpClient.newCall(request).execute()
+            val inputStream = BufferedInputStream(response.body.byteStream())
+            val bufferedReader = BufferedReader(InputStreamReader(inputStream))
+            var line: String? = bufferedReader.readLine()
+            while (line != null) {
+                out.append(line)
+                line = bufferedReader.readLine()
+            }
+            bufferedReader.close()
+        } catch (e: Exception) {
+            UtilityLog.handleException(e)
+        }
+        return out.toString()
+    }
+
+    // PolygonWarning.kt
+    // FYI - this is probably not needed and could use getStringFromUrlBaseNoAcceptHeader1 instead
+    fun getStringFromUrlBaseNoHeader1(url: String): String {
+        UtilityLog.download("getStringFromUrlBaseNoHeader1: $url")
+        val out = StringBuilder(5000)
+        try {
+            val request = Request.Builder()
+                .url(url)
+                .header("User-Agent", USER_AGENT)
+                .addHeader("Accept", ACCEPT_HEADER)
+                .build()
+            val response = MyApplication.httpClient.newCall(request).execute()
+            val inputStream = BufferedInputStream(response.body.byteStream())
+            val bufferedReader = BufferedReader(InputStreamReader(inputStream))
+            var line: String? = bufferedReader.readLine()
+            while (line != null) {
+                out.append(line)
+                line = bufferedReader.readLine()
+            }
+            bufferedReader.close()
+        } catch (e: Exception) {
+            UtilityLog.handleException(e)
+        }
+        return out.toString()
+    }
+
+    // used by CapAlert.kt
+    fun getStringFromUrlSep(url: String): String {
+        UtilityLog.download("getStringFromUrlSep: $url")
+        val breakStr = "ABC123_456ZZ"
+        val out = StringBuilder(5000)
+        val request = Request.Builder()
+            .url(url)
+            .header("User-Agent", USER_AGENT)
+            .addHeader(
+                "Accept",
+                "application/vnd.noaa.dwml+xml;version=1"
+            ).build()
+        requestToString(request, breakStr, out)
+        return out.toString().replace(breakStr, "<br>")
+    }
+
+//    fun getStringFromUrlSep(url: String): String {
+//        UtilityLog.download("getStringFromUrlSep: $url")
+//        val breakStr = "ABC123_456ZZ"
+//        val out = StringBuilder(5000)
+//        try {
+//            val request = Request.Builder()
+//                .url(url)
+//                .header("User-Agent", USER_AGENT)
+//                .addHeader(
+//                    "Accept",
+//                    "application/vnd.noaa.dwml+xml;version=1"
+//                ) // TODO FIXME, not valid defaulting to application/geo+json
+//                .build()
+//            val response = MyApplication.httpClient.newCall(request).execute()
+//            val inputStream = BufferedInputStream(response.body.byteStream())
+//            val bufferedReader = BufferedReader(InputStreamReader(inputStream))
+//            var line: String? = bufferedReader.readLine()
+//            while (line != null) {
+//                out.append(line)
+//                line = bufferedReader.readLine()
+//            }
+//            out.append(breakStr)
+//            bufferedReader.close()
+//        } catch (e: Exception) {
+//            UtilityLog.handleException(e)
+//        }
+//        return out.toString().replace(breakStr, "<br>")
+//    }
+
+    private fun requestToString(request: Request, breakString: String, output: StringBuilder) {
+        try {
+            val response = MyApplication.httpClient.newCall(request).execute()
+            val inputStream = BufferedInputStream(response.body.byteStream())
+            val bufferedReader = BufferedReader(InputStreamReader(inputStream))
+            var line: String? = bufferedReader.readLine()
+            while (line != null) {
+                output.append(line)
+                line = bufferedReader.readLine()
+            }
+            output.append(breakString)
+            bufferedReader.close()
+        } catch (e: Exception) {
+            UtilityLog.handleException(e)
+        }
     }
 }